Patch for dcm4jboss v2.2.1 to remove DB record before file deletion on file system

+++ dcm4chee/dcm4chee-arc/branches/DCM4JBOSS_2_2_1_BRANCH/dcm4jboss-sar/src/java/org/dcm4chex/archive/mbean/CompressionService.java
@@ -34,6 +34,7 @@ import org.dcm4chex.archive.config.RetryIntervalls;
 import org.dcm4chex.archive.ejb.interfaces.FileDTO;
 import org.dcm4chex.archive.ejb.interfaces.FileSystemMgt;
 import org.dcm4chex.archive.ejb.interfaces.FileSystemMgtHome;
+import org.dcm4chex.archive.mbean.CompressionService.CompressionRule;
 import org.dcm4chex.archive.util.EJBHomeFactory;
 import org.dcm4chex.archive.util.FileUtils;
 
@@ -252,17 +253,19 @@ public class CompressionService extends TimerSupport {
         }
     }
 
-     private void doCompress(FileSystemMgt fsMgt, FileDTO fileDTO,
+    private void doCompress(FileSystemMgt fsMgt, FileDTO fileDTO,
             CompressionRule info) {
         File baseDir = FileUtils.toFile(fileDTO.getDirectoryPath());
         File srcFile = FileUtils.toFile(fileDTO.getDirectoryPath(), fileDTO
                 .getFilePath());
-        File destFile = incFileName(srcFile);
-        if (log.isDebugEnabled())
-            log.debug("Compress file " + srcFile + " to " + destFile
-                    + " with CODEC:" + info.getCodec() + "("
-                    + info.getTransferSyntax() + ")");
+		File destFile = null;
         try {
+	        destFile = FileUtils.createNewFile(srcFile.getParentFile(),
+					(int) Long.parseLong(srcFile.getName(), 16)+1);
+	        if (log.isDebugEnabled())
+	            log.debug("Compress file " + srcFile + " to " + destFile
+	                    + " with CODEC:" + info.getCodec() + "("
+	                    + info.getTransferSyntax() + ")");
             int[] pxvalVR = new int[1];
             byte[] md5 = CompressCmd.compressFile(srcFile, destFile, info
                     .getTransferSyntax(), pxvalVR);
@@ -298,10 +301,10 @@ public class CompressionService extends TimerSupport {
             if (log.isDebugEnabled())
                 log.debug("replace File " + srcFile + " with " + destFile);
             fsMgt.replaceFile(fileDTO.getPk(), destFilePath, info
-                    .getTransferSyntax(), (int) srcFile.length(), md5);
+                    .getTransferSyntax(), (int) destFile.length(), md5);
         } catch (Exception x) {
             log.error("Can't compress file:" + srcFile, x);
-            if (destFile.exists())
+            if (destFile != null && destFile.exists())
                 destFile.delete();
             try {
                 fsMgt.setFileStatus(fileDTO.getPk(), FileDTO.COMPRESS_FAILED);
@@ -310,16 +313,6 @@ public class CompressionService extends TimerSupport {
                         + srcFile);
             }
         }
-    }
-
-    private File incFileName(File src) {
-        File path = src.getParentFile();
-        long fnAsInt = Long.parseLong(src.getName(), 16);
-        File f;
-        do {
-            f = new File(path, Long.toHexString(++fnAsInt).toUpperCase());            
-        } while (f.exists());
-        return f;
     }
 
     private boolean isDisabled(int hour) {
